Fix race condition
If a point is being processed in one go routine and another comes, it dropped the queue entry. However, if the position has become accessible AFTER the current processing routine has checked whether the roll can be removed, the roll was not removed and the second check silently dropped. Instead of dropping in progress rolls, we requeue them.